package_file {pkgload}	R Documentation

Find file in a package.

Description

It always starts by finding by walking up the path until it finds the root directory, i.e. a directory containing DESCRIPTION. If it cannot find the root directory, or it can't find the specified path, it will throw an error.

Usage

package_file(..., path = ".")

Arguments

...
Components of the path.

path
Place to start search for package directory.

Examples

## Not run: 
package_file("figures", "figure_1")

## End(Not run)

[Package pkgload version 1.3.3 Index]