packages {pkgload}	R Documentation

Helper functions for working with development packages.

Description

All functions search recursively up the directory tree from the input path until they find a DESCRIPTION file.

Usage

pkg_path(path = ".")

pkg_name(path = ".")

pkg_desc(path = ".")

pkg_version(path = ".")

pkg_ns(path = ".")

Arguments

path
Path to a package, or within a package.

Functions

pkg_path: Return the normalized package path.

pkg_name: Return the package name.

pkg_desc: Return the package DESCRIPTION as a desc::desc() object.

pkg_version: Return the package version.

pkg_ns: Return the package namespace.

[Package pkgload version 1.2.4 Index]