pkgload-package {pkgload}	R Documentation

pkgload: Simulate Package Installation and Attach

Description

Simulates the process of installing a package and then attaching it. This is a key part of the 'devtools' package as it allows you to rapidly iterate while developing a package.

Author(s)

Maintainer: Lionel Henry lionel@rstudio.com

Authors:

Hadley Wickham

Winston Chang

Jim Hester

Other contributors:

RStudio [copyright holder, funder]

R Core team (Some namespace and vignette code extracted from base R) [contributor]

See Also

Useful links:

https://github.com/r-lib/pkgload

https://pkgload.r-lib.org

Report bugs at https://github.com/r-lib/pkgload/issues

[Package pkgload version 1.3.1 Index]