check_dep_version {pkgload}	R Documentation

Check that the version of an imported package satisfies the requirements

Description

Check that the version of an imported package satisfies the requirements

Usage

check_dep_version(dep_name, dep_ver = "*")

Arguments

dep_name
The name of the package with objects to import

dep_ver
The version of the package, this includes the specified comparison operator.

[Package pkgload version 1.2.4 Index]