imports_env {pkgload}	R Documentation

Return imports environment for a package

Description

Contains objects imported from other packages. Is the parent of the package namespace environment, and is a child of <namespace:base>, which is a child of R_GlobalEnv.

Usage

imports_env(package)

Arguments

path
TODO: fix doc

See Also

ns_env() for the namespace environment that all the objects (exported and not exported).

pkg_env() for the attached environment that contains the exported objects.

[Package pkgload version 1.2.4 Index]